refactor: reimplement audio alignment with a new service architecture
