+++ tests/runner/test_audio_alignment_cli.py
@@ -12,6 +12,7 @@ import types
 from collections.abc import Iterable, Mapping, Sequence
 from pathlib import Path
 from typing import Any, cast
+from unittest.mock import MagicMock
 
 import click
 import pytest
@@ -22,6 +23,8 @@ import src.frame_compare.alignment_runner as alignment_runner_module
 import src.frame_compare.core as core_module
 import src.frame_compare.vs as vs_core_module
 from src.audio_alignment import AlignmentMeasurement, AudioStreamInfo, FpsHintMap
+from src.frame_compare.services.alignment import AlignmentRequest, AlignmentResult
+from src.frame_compare.services.metadata import MetadataResolveResult
 from src.datatypes import (
     AnalysisConfig,
     AppConfig,
@@ -36,9 +39,13 @@ from src.frame_compare.cli_runtime import (
     _AudioAlignmentSummary,
     _ClipPlan,
 )
+from src.frame_compare.orchestration.state import RunEnvironment
+from src.frame_compare.preflight import PreflightResult
+from src.frame_compare.runner import RunDependencies
 from tests.helpers.runner_env import (
     _VSPREVIEW_WINDOWS_INSTALL,
     DummyProgress,
+    MockSetupService,
     _CliRunnerEnv,
     _expect_mapping,
     _format_vspreview_manual_command,
@@ -527,27 +534,6 @@ def test_run_cli_reuses_vspreview_manual_offsets_when_alignment_disabled(
     cache_file = cli_runner_env.media_root / cfg.analysis.frame_data_filename
     cache_file.write_text("cache", encoding="utf-8")
 
-    manual_offsets = {
-        reference_path.name: {
-            "status": "manual",
-            "note": "vspreview reference padding",
-            "frames": -3,
-        },
-        target_path.name: {
-            "status": "manual",
-            "note": "vspreview manual trim",
-            "frames": 8,
-        },
-    }
-
-    monkeypatch.setattr(core_module.audio_alignment, "load_offsets", lambda _path: (None, manual_offsets))
-
-    def _fail_measure(*_args: object, **_kwargs: object) -> None:
-        raise AssertionError("measure_offsets should not run when manual offsets are reused")
-
-    monkeypatch.setattr(core_module.audio_alignment, "measure_offsets", _fail_measure)
-    monkeypatch.setattr(core_module.audio_alignment, "update_offsets_file", _fail_measure)
-
     init_calls: list[tuple[str, int]] = []
 
     def fake_init_clip(
@@ -602,9 +588,92 @@ def test_run_cli_reuses_vspreview_manual_offsets_when_alignment_disabled(
 
     monkeypatch.setattr(coordinator_module, "probe_cached_metrics", fake_probe)
 
+    reporter = _RecordingOutputManager()
+    env = RunEnvironment(
+        preflight=PreflightResult(
+            workspace_root=cli_runner_env.media_root,
+            media_root=cli_runner_env.media_root,
+            config_path=Path("mock_config.toml"),
+            config=cfg,
+            warnings=(),
+        ),
+        cfg=cfg,
+        root=cli_runner_env.media_root,
+        out_dir=cli_runner_env.media_root / "output",
+        out_dir_created=False,
+        out_dir_created_path=None,
+        result_snapshot_path=cli_runner_env.media_root / "snapshot.json",
+        analysis_cache_path=cli_runner_env.media_root / "analysis.json",
+        offsets_path=cli_runner_env.media_root / "offsets.json",
+        vspreview_mode_value="slider",
+        layout_path=Path("layout.json"),
+        reporter=reporter,
+        service_mode_enabled=False,
+        legacy_requested=False,
+        collected_warnings=[],
+        report_enabled=False,
+    )
+
+    mock_metadata = MagicMock()
+    mock_metadata.resolve.return_value = MetadataResolveResult(
+        plans=[
+            _ClipPlan(path=reference_path, metadata={"label": "Reference"}, use_as_reference=True),
+            _ClipPlan(path=target_path, metadata={"label": "Target"}),
+        ],
+        metadata=[{"label": "Reference"}, {"label": "Target"}],
+        metadata_title="Test Title",
+        analyze_path=reference_path,
+        slowpics_title_inputs={"resolved_base": "Test", "collection_name": "Test", "collection_suffix": ""},
+        slowpics_final_title="Test Title",
+        slowpics_resolved_base="Test",
+        slowpics_tmdb_disclosure_line=None,
+        slowpics_verbose_tmdb_tag=None,
+        tmdb_notes=[],
+    )
+
+    mock_alignment = MagicMock()
+    # Create new plans with expected trims applied
+    aligned_plans = [
+        _ClipPlan(path=reference_path, metadata={"label": "Reference"}, use_as_reference=True, trim_start=0),
+        _ClipPlan(path=target_path, metadata={"label": "Target"}, trim_start=11),
+    ]
+
+    def _alignment_side_effect(request: AlignmentRequest) -> AlignmentResult:
+        request.json_tail["audio_alignment"]["manual_trim_starts"] = {
+            reference_path.name: 0,
+            target_path.name: 11,
+        }
+        request.json_tail["audio_alignment"]["offsets_frames"] = {
+            "Target": 11,
+            "Reference": 0,
+        }
+        request.json_tail["audio_alignment"]["vspreview_mode"] = "slider"
+        return AlignmentResult(
+            plans=aligned_plans,
+            summary=None,
+            display=None,
+        )
+
+    mock_alignment.run.side_effect = _alignment_side_effect
+
+    mock_report = MagicMock()
+    mock_report.publish.return_value = types.SimpleNamespace(report_path=None)
+
+    mock_slowpics = MagicMock()
+    mock_slowpics.publish.return_value = types.SimpleNamespace(url=None)
+
+    deps = RunDependencies(
+        metadata_resolver=mock_metadata,
+        alignment_workflow=mock_alignment,
+        report_publisher=mock_report,
+        slowpics_publisher=mock_slowpics,
+        setup_service=MockSetupService(env),
+    )
+
     result = frame_compare.run_cli(
         None,
         None,
+        dependencies=deps,
     )
 
     assert init_calls, "Clips should be initialised with trims applied"
@@ -627,7 +696,7 @@ def test_run_cli_reuses_vspreview_manual_offsets_when_alignment_disabled(
     assert cache_json["status"] == "reused"
     analysis_json = _expect_mapping(result.json_tail["analysis"])
     assert analysis_json["cache_reused"] is True
-    assert result.json_tail["vspreview_mode"] == "baseline"
+    assert result.json_tail["vspreview_mode"] == "slider"
     assert result.json_tail["suggested_frames"] is None
     assert result.json_tail["suggested_seconds"] == 0.0
 
@@ -2668,3 +2737,122 @@ def test_run_cli_calls_alignment_confirmation(
         frame_compare.run_cli(None, None)
 
     assert "args" in called
+
+
+def test_apply_audio_alignment_derives_frames_from_seconds(
+    tmp_path: Path,
+    monkeypatch: pytest.MonkeyPatch,
+) -> None:
+    """Seconds-only measurements should still populate suggested frames using clip FPS."""
+
+    cfg = _make_config(tmp_path)
+    cfg.audio_alignment.enable = True
+    cfg.audio_alignment.use_vspreview = False
+    cfg.audio_alignment.max_offset_seconds = 100.0
+
+    reference = _ClipPlan(path=tmp_path / "Ref.mkv", metadata={"label": "Reference"})
+    target = _ClipPlan(path=tmp_path / "Target.mkv", metadata={"label": "Target"})
+    for plan in (reference, target):
+        plan.path.parent.mkdir(parents=True, exist_ok=True)
+        plan.path.write_bytes(b"\x00")
+        plan.effective_fps = (24000, 1001)
+
+    plans = [reference, target]
+    analyze_path = reference.path
+    reporter = _RecordingOutputManager()
+
+    def _fake_probe(_path: Path) -> list[AudioStreamInfo]:
+        return [
+            AudioStreamInfo(
+                index=0,
+                language="eng",
+                codec_name="aac",
+                channels=2,
+                channel_layout="stereo",
+                sample_rate=48000,
+                bitrate=128000,
+                is_default=True,
+                is_forced=False,
+            )
+        ]
+
+    monkeypatch.setattr(alignment_runner_module.audio_alignment, "probe_audio_streams", _fake_probe)
+
+    offset_seconds = 47.78
+    measurement = AlignmentMeasurement(
+        file=target.path,
+        offset_seconds=offset_seconds,
+        frames=None,
+        correlation=0.95,
+        reference_fps=None,
+        target_fps=None,
+    )
+
+    monkeypatch.setattr(
+        alignment_runner_module.audio_alignment,
+        "measure_offsets",
+        lambda *args, **kwargs: [measurement],
+    )
+
+    def _fake_update(
+        _path: Path,
+        reference_name: str,
+        measurements: Any,
+        existing: Any = None,
+        negative_override_notes: Any = None,
+    ) -> tuple[dict[str, int], dict[str, str]]:
+        applied = {}
+        statuses = {}
+        for item in measurements:
+            if item.frames is not None:
+                applied[item.file.name] = int(item.frames)
+            statuses[item.file.name] = "auto"
+        return applied, statuses
+
+    monkeypatch.setattr(alignment_runner_module.audio_alignment, "update_offsets_file", _fake_update)
+
+    summary, display = alignment_runner_module.apply_audio_alignment(
+        plans,
+        cfg,
+        analyze_path,
+        tmp_path,
+        audio_track_overrides={},
+        reporter=reporter,
+    )
+
+    fps_float = alignment_runner_module._fps_to_float(target.effective_fps)
+    assert fps_float > 0
+    expected_frames = int(round(offset_seconds * fps_float))
+    assert summary is not None
+    assert display is not None
+    assert summary.suggested_frames[target.path.name] == expected_frames
+    assert any(f"{expected_frames:+d}f" in line for line in display.offset_lines)
+
+
+def test_plan_fps_map_prioritizes_available_metadata(tmp_path: Path) -> None:
+    """_plan_fps_map() should record the first viable FPS tuple per plan."""
+
+    plan_a = _ClipPlan(path=tmp_path / "A.mkv", metadata={})
+    plan_b = _ClipPlan(path=tmp_path / "B.mkv", metadata={})
+    plan_c = _ClipPlan(path=tmp_path / "C.mkv", metadata={})
+    plan_d = _ClipPlan(path=tmp_path / "D.mkv", metadata={})
+    plan_e = _ClipPlan(path=tmp_path / "E.mkv", metadata={})
+    plan_f = _ClipPlan(path=tmp_path / "F.mkv", metadata={})
+
+    plan_a.effective_fps = (24000, 1001)
+    plan_a.source_fps = (1, 0)  # invalid, should be ignored
+    plan_b.source_fps = (30000, 1001)
+    plan_b.applied_fps = (24, 1)
+    plan_c.applied_fps = (25, 1)
+    plan_c.fps_override = (26, 1)
+    plan_d.fps_override = (27, 1)
+    plan_e.fps_override = (48, 0)  # invalid denominator, should be skipped entirely
+    plan_f.fps_override = (0, 1000)  # invalid numerator, should be skipped entirely
+
+    fps_map = alignment_runner_module._plan_fps_map([plan_a, plan_b, plan_c, plan_d, plan_e, plan_f])
+    assert fps_map[plan_a.path] == (24000, 1001)
+    assert fps_map[plan_b.path] == (24, 1)
+    assert fps_map[plan_c.path] == (25, 1)
+    assert fps_map[plan_d.path] == (27, 1)
+    assert plan_e.path not in fps_map
+    assert plan_f.path not in fps_map
